Add Key stage information to HTB Template object
