+++ ApplyToBecomeInternal/ApplyToBecomeInternal/ViewModels/KeyStage2PerformanceTableViewModel.cs
@@ -1,4 +1,8 @@
-﻿using Microsoft.AspNetCore.Html;
+using ApplyToBecome.Data.Models.KeyStagePerformance;
+using ApplyToBecomeInternal.Extensions;
+using static ApplyToBecomeInternal.Extensions.DisplayExtensions;
+
+using Microsoft.AspNetCore.Html;
 
 namespace ApplyToBecomeInternal.ViewModels
 {
@@ -20,5 +24,28 @@ namespace ApplyToBecomeInternal.ViewModels
 		public string LAAverageReadingProgressScore { get; set; }
 		public string LAAverageWritingProgressScore { get; set; }
 		public string LAAverageMathsProgressScore { get; set; }
+
+		public static KeyStage2PerformanceTableViewModel Build(KeyStage2PerformanceResponse keyStage2Performance)
+		{
+			return new KeyStage2PerformanceTableViewModel
+			{
+				Year = keyStage2Performance.Year.FormatKeyStageYear(),
+				PercentageMeetingExpectedStdInRWM = keyStage2Performance.PercentageMeetingExpectedStdInRWM.NotDisadvantaged.FormatValue(),
+				PercentageAchievingHigherStdInRWM = keyStage2Performance.PercentageAchievingHigherStdInRWM.NotDisadvantaged.FormatValue(),
+				ReadingProgressScore = keyStage2Performance.ReadingProgressScore.NotDisadvantaged.FormatValue(),
+				WritingProgressScore = keyStage2Performance.WritingProgressScore.NotDisadvantaged.FormatValue(),
+				MathsProgressScore = keyStage2Performance.MathsProgressScore.NotDisadvantaged.FormatValue(),
+				NationalAveragePercentageMeetingExpectedStdInRWM = HtmlFormatKeyStageDisadvantagedResult(keyStage2Performance.NationalAveragePercentageMeetingExpectedStdInRWM),
+				NationalAveragePercentageAchievingHigherStdInRWM = HtmlFormatKeyStageDisadvantagedResult(keyStage2Performance.NationalAveragePercentageAchievingHigherStdInRWM),
+				NationalAverageReadingProgressScore = keyStage2Performance.NationalAverageReadingProgressScore.NotDisadvantaged.FormatValue(),
+				NationalAverageWritingProgressScore = keyStage2Performance.NationalAverageWritingProgressScore.NotDisadvantaged.FormatValue(),
+				NationalAverageMathsProgressScore = keyStage2Performance.NationalAverageMathsProgressScore.NotDisadvantaged.FormatValue(),
+				LAAveragePercentageMeetingExpectedStdInRWM = keyStage2Performance.LAAveragePercentageMeetingExpectedStdInRWM.NotDisadvantaged.FormatValue(),
+				LAAveragePercentageAchievingHigherStdInRWM = keyStage2Performance.LAAveragePercentageAchievingHigherStdInRWM.NotDisadvantaged.FormatValue(),
+				LAAverageMathsProgressScore = keyStage2Performance.LAAverageMathsProgressScore.NotDisadvantaged.FormatValue(),
+				LAAverageReadingProgressScore = keyStage2Performance.LAAverageReadingProgressScore.NotDisadvantaged.FormatValue(),
+				LAAverageWritingProgressScore = keyStage2Performance.LAAverageWritingProgressScore.NotDisadvantaged.FormatValue()
+			};
+		}
 	}
 }
